+++ tools/level_editor/project_browser.vala
@@ -75,10 +75,16 @@ public GLib.Menu? project_entry_menu_common(ProjectStore.RowKind kind, string ty
 }
 
 // Menu to open when clicking on project's files and folders.
-public GLib.Menu? project_entry_menu_create(ProjectStore.RowKind kind, string type, string name)
+public GLib.Menu? project_entry_menu_create(BrowseMode mode, ProjectStore.RowKind kind, string type, string name)
 {
 	GLib.Menu menu = new GLib.Menu();
 	GLib.MenuItem mi;
+
+	if (mode == BrowseMode.SEARCH && (kind == ProjectStore.RowKind.FOLDER || kind == ProjectStore.RowKind.RESOURCE)) {
+		mi = new GLib.MenuItem(_("Reveal"), null);
+		mi.set_action_and_target_value("app.reveal-resource", new GLib.Variant.tuple({new GLib.Variant.int32((int)kind), type, name}));
+		menu.append_item(mi);
+	}
 
 	if (kind == ProjectStore.RowKind.FOLDER) {
 		if (name == "..")
@@ -132,8 +138,6 @@ public GLib.Menu? project_entry_menu_create(ProjectStore.RowKind kind, string ty
 
 		menu.append_section(null, destroy_menu);
 	} else { // If file
-		menu = new GLib.Menu();
-
 		if (type != "lua") {
 			mi = new GLib.MenuItem(_("Duplicate..."), null);
 			mi.set_action_and_target_value("app.duplicate-resource", new GLib.Variant.string(project_path(kind, type, name)));
@@ -804,7 +808,7 @@ public class ProjectFolderView : Gtk.Box
 
 			GLib.Menu? menu_model;
 			if (_showing_project_folder)
-				menu_model = project_entry_menu_create(kind, type, name);
+				menu_model = project_entry_menu_create(_browse_mode, kind, type, name);
 			else
 				menu_model = favorites_entry_menu_create(kind, type, name);
 
@@ -2431,7 +2435,7 @@ public class ProjectBrowser : Gtk.Box
 			Gtk.TreePath? store_path = _tree_filter.convert_path_to_child_path(search_path);
 			GLib.Menu? menu_model;
 			if (store_path.is_descendant(_project_store.project_root_path()) || store_path.compare(_project_store.project_root_path()) == 0)
-				menu_model = project_entry_menu_create((ProjectStore.RowKind)kind, (string)type, (string)name);
+				menu_model = project_entry_menu_create(_browse_mode, (ProjectStore.RowKind)kind, (string)type, (string)name);
 			else if (store_path.is_descendant(_project_store.favorites_root_path()))
 				menu_model = favorites_entry_menu_create((ProjectStore.RowKind)kind, (string)type, (string)name);
 			else